+++ app/serializers/person/duplicate_serializer.rb
@@ -1,6 +1,0 @@
-class Person::DuplicateSerializer < ActiveModel::Serializer
-  has_one :person
-  has_one :dup_person
-  has_one :shared_contact
-  has_many :people
-end